pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "networkii"
version = "1.0.0"
description = "Network monitoring and statistics tool for Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "pillow",
    "spidev",
    "displayhatmini",
    "netifaces",
    "numpy",
    "speedtest-cli",
    "rpi-lgpio",
    "rich",
]

[project.scripts]
networkii = "networkii.cli.networkii:main"

[tool.hatch.build.targets.wheel]
packages = ["networkii"] 